Clean up ast.py
